refactor: 값 타입 패키지 수정
