+++ src/main/java/domain/LadderFactory.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.value.Height;
+import domain.value.Width;
+
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
