+++ src/main/java/domain/value/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.value;
 
 public class Name {
