+++ src/main/java/domain/value/WinningEntry.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.value;
 
 public class WinningEntry {
