+++ src/test/java/domain/value/HeightTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.value;
 
+import domain.value.Height;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -11,14 +12,14 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@DisplayName("Width 는")
-class WidthTest {
+@DisplayName("Height 는")
+class HeightTest {
 
-    @ParameterizedTest(name = "너비는 1 이상이어야 한다")
+    @ParameterizedTest(name = "높이는 1 이상이어야 한다")
     @ValueSource(ints = {-100, -1, 0})
-    void 너비는_1_이상이어야_한다(final int notNaturalNumber) {
+    void 높이는_1_이상이어야_한다(final int notNaturalNumber) {
         // when & then
-        assertThatThrownBy(() -> Width.of(notNaturalNumber))
+        assertThatThrownBy(() -> Height.of(notNaturalNumber))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -26,10 +27,10 @@ class WidthTest {
     @ValueSource(ints = {1, 10, 50, 100})
     void 생성될_때_전달받은_값을_가진다(final int number) {
         // given
-        Width width = Width.of(number);
+        Height height = Height.of(number);
 
         // when
-        int value = width.value();
+        int value = height.value();
 
         // then
         assertThat(value).isEqualTo(number);
@@ -39,10 +40,10 @@ class WidthTest {
     @ValueSource(ints = {1, 10, 50, 100})
     void 값이_같으면_동등하다(final int number) {
         // given
-        Width width1 = Width.of(number);
-        Width width2 = Width.of(number);
+        Height height1 = Height.of(number);
+        Height height2 = Height.of(number);
 
         // when & then
-        assertThat(width1).isEqualTo(width2);
+        assertThat(height1).isEqualTo(height2);
     }
 }
