+++ src/test/java/domain/value/NamesTest.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.value;
 
+import domain.value.Name;
+import domain.value.Names;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
