+++ src/test/java/domain/value/WinningEntryTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.value;
 
+import domain.value.WinningEntry;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
